+++ e2e/test-environment-async/TestEnvironment.js
@@ -10,11 +10,13 @@ const JSDOMEnvironment = require('jest-environment-jsdom');
 const DIR = os.tmpdir() + '/jest-test-environment';
 
 class TestEnvironment extends JSDOMEnvironment {
-  constructor(config) {
-    super(config);
+  constructor(config, context) {
+    super(config, context);
+    this.context = context;
   }
 
   setup() {
+    console.info('TestEnvironment.setup:', this.context.testPath);
     return super.setup().then(() => {
       this.global.setup = 'setup';
     });